fix: also use long long in bindings

+++ k2.h
@@ -1,7 +1,7 @@
 #ifndef K2_H
 #define K2_H
 
-extern "C" void k2_params(unsigned long chunkseed, const int *neighbors, int diagonal, int height);
-extern "C" unsigned long k2_start(int threads, unsigned long begin, unsigned long end, unsigned long *out);
+extern "C" void k2_params(unsigned long long chunkseed, const int *neighbors, int diagonal, int height);
+extern "C" unsigned long long k2_start(int threads, unsigned long long begin, unsigned long long end, unsigned long long *out);
 
 #endif
